Added support for deflated responses.
I ran into this when using dghubble/go-twitter and hit the rate
limit. Funnily enough, the response from the twitter api contained
the Content-Encoding: deflate header value. Trying to unmarshal
the compressed bytes resulted in a json syntax error:

    invalid character 'x' looking for beginning of value

Which is super confusing. It took me a while to look at the
response and see the statuscode was 429. Anyways, it seems to work
now, and when hitting the twitter rate limit it parses nicely to
failureV.